# [rust] Batch nearby HTTP requests for bbox features