~~DRAFT: because it's based on #318 . Please review that first.~~ Merged!

When reading FGB over HTTP, most time is spent in the network, as opposed to local processing.

To minimize network round trips, we currently "over fetch" whenever we're getting feature data, in hopes of obviating a future request. For "select_all" queries this works great, because eventually we're going to need all the data anyway.

For bbox queries, we specifically don't want *all* the data.  Thanks to hilbert ordering, spatially nearby features are typically nearby on disk so we do OK with the same strategy, but inevitably we end up fetching data we don't end up using. 

The hilbert index contains the start location of every feature. We can infer the length of each feature by knowing where the _next_ feature starts. So we actually have all the information to fetch precisely the feature data we need for a given bbox.

So we could precisely fetch every feature individually to minimize bandwidth, but because network round trips are expensive, it still makes sense to "waste" some bytes and combine these feature requests. But now we can quantify exactly how many bytes we're willing to waste to obviate a separate request, and we don't bother fetching extra data when we know it's not useful.

I made a similar PR for JS/TS a while back (though that one also does concurrent fetching): https://github.com/flatgeobuf/flatgeobuf/pull/135

## Time improvements

Recall that our benchmarks are run against a simulated network which has 100ms latency and 50Mbps bandwidth.

```
medium select_bbox
time:   [636.61 ms 637.89 ms 639.27 ms]
change: [-41.733% -41.515% -41.296%] (p = 0.00 < 0.05)
Performance has improved.

medium bbox -> geojson
time:   [661.92 ms 663.29 ms 664.79 ms]
change: [-40.336% -40.172% -39.995%] (p = 0.00 < 0.05)
Performance has improved.
```

The rest of the benches are unaffected. Since `select_all`'s fetch strategy didn't change and bbox requests on small files already fit into the over-fetch of our initial header request, this is exactly as expected.

## Network improvements

In the above example, byte consumption was reduced from 4.25MB to 1.78MB

Before feature batching:
```
req_count: 6, body_bytes_sent: 4253968
```

With feature batching:
```
req_count: 6, body_bytes_sent: 1775808
```

Note that, in this particular case from our existing benches, the total number of requests were not reduced. But if the feature batches are close enough, and we're within our maximum request size, we will sometimes consolidate requests.

Note we still want to have _some_ maximum request size, because our current chunking approach stores the entire response in memory.